+++ tests/opda/test_parametric.py
@@ -863,6 +863,120 @@ class QuadraticDistributionTestCase(testcases.RandomTestCase):
                     self.assertGreaterEqual(c, bounds[2, 0])
                     self.assertLessEqual(c, bounds[2, 1])
 
+        # Test error conditions.
+        for convex in [False, True]:
+            # when ys is not 1D
+            with self.assertRaises(ValueError):
+                parametric.QuadraticDistribution\
+                    .estimate_initial_parameters_and_bounds(
+                        0,
+                        fraction=0.5,
+                        convex=convex,
+                    )
+            with self.assertRaises(ValueError):
+                parametric.QuadraticDistribution\
+                    .estimate_initial_parameters_and_bounds(
+                        [[0]],
+                        fraction=0.5,
+                        convex=convex,
+                    )
+            with self.assertRaises(ValueError):
+                parametric.QuadraticDistribution\
+                    .estimate_initial_parameters_and_bounds(
+                        [[0], [1]],
+                        fraction=0.5,
+                        convex=convex,
+                    )
+            # when ys is empty
+            with self.assertRaises(ValueError):
+                parametric.QuadraticDistribution\
+                    .estimate_initial_parameters_and_bounds(
+                        [],
+                        fraction=0.5,
+                        convex=convex,
+                    )
+            # when fraction is not a scalar
+            with self.assertRaises(ValueError):
+                parametric.QuadraticDistribution\
+                    .estimate_initial_parameters_and_bounds(
+                        [0, 1, 2, 3, 4, 5],
+                        fraction=[0.5],
+                        convex=convex,
+                    )
+            with self.assertRaises(ValueError):
+                parametric.QuadraticDistribution\
+                    .estimate_initial_parameters_and_bounds(
+                        [0, 1, 2, 3, 4, 5],
+                        fraction=[[0.5]],
+                        convex=convex,
+                    )
+            # when fraction is not between 0 and 1, inclusive of 1
+            with self.assertRaises(ValueError):
+                parametric.QuadraticDistribution\
+                    .estimate_initial_parameters_and_bounds(
+                        [0, 1, 2, 3, 4, 5],
+                        fraction=-0.1,
+                        convex=convex,
+                    )
+            with self.assertRaises(ValueError):
+                parametric.QuadraticDistribution\
+                    .estimate_initial_parameters_and_bounds(
+                        [0, 1, 2, 3, 4, 5],
+                        fraction=0.,
+                        convex=convex,
+                    )
+            with self.assertRaises(ValueError):
+                parametric.QuadraticDistribution\
+                    .estimate_initial_parameters_and_bounds(
+                        [0, 1, 2, 3, 4, 5],
+                        fraction=1.1,
+                        convex=convex,
+                    )
+            # when convex is not a bool
+            with self.assertRaises(TypeError):
+                parametric.QuadraticDistribution\
+                    .estimate_initial_parameters_and_bounds(
+                        [0, 1, 2, 3, 4, 5],
+                        fraction=0.5,
+                        convex=None,
+                    )
+            with self.assertRaises(TypeError):
+                parametric.QuadraticDistribution\
+                    .estimate_initial_parameters_and_bounds(
+                        [0, 1, 2, 3, 4, 5],
+                        fraction=0.5,
+                        convex=0.,
+                    )
+            with self.assertRaises(TypeError):
+                parametric.QuadraticDistribution\
+                    .estimate_initial_parameters_and_bounds(
+                        [0, 1, 2, 3, 4, 5],
+                        fraction=0.5,
+                        convex=[True],
+                    )
+            # when fraction * len(ys) < 1
+            with self.assertRaises(ValueError):
+                parametric.QuadraticDistribution\
+                    .estimate_initial_parameters_and_bounds(
+                        [0],
+                        fraction=0.99,
+                        convex=convex,
+                    )
+            with self.assertRaises(ValueError):
+                parametric.QuadraticDistribution\
+                    .estimate_initial_parameters_and_bounds(
+                        [0, 1],
+                        fraction=0.49,
+                        convex=convex,
+                    )
+            with self.assertRaises(ValueError):
+                parametric.QuadraticDistribution\
+                    .estimate_initial_parameters_and_bounds(
+                        [0, 1, 2],
+                        fraction=0.32,
+                        convex=convex,
+                    )
+
     def test_sample_defaults_to_global_random_number_generator(self):
         # sample should be deterministic if global seed is set.
         dist = parametric.QuadraticDistribution(0., 1., 2)